Added ttl file for finance goals
I created a simple ttl file for the finance goals to save basic goal information to the solid pod. The work on the goals section is intended to be a basic example and should not impact the functionality of any of the features. The goals section is sort of acting as a sandbox for getting database functionality working.